infra/lambda/api-authorizer/requirements.txt
python-jose[cryptography]>=3.3.0
